Fix clippy warnings in linux.rs

+++ src/linux.rs
@@ -17,7 +17,7 @@ pub async fn get_info() -> Result<ProcessStats, Error> {
     let stat_file =
         Stat::from_reader(readable_string).map_err(|_e| Error::FileContentsMalformed)?;
 
-    let memory_usage_bytes = (stat_file.rss as u64) * (bytes_per_page as u64);
+    let memory_usage_bytes = stat_file.rss * bytes_per_page; // both are u64
     let user_mode_seconds = (stat_file.utime as f64) / (ticks_per_second as f64);
     let kernel_mode_seconds = (stat_file.stime as f64) / (ticks_per_second as f64);
